+++ mod/luaobjectinfo.lua
@@ -1,3 +1,4 @@
+local invert = require("__debugadapter__/enumutil.lua").invert
 local __DebugAdapter = __DebugAdapter
 local pcall = pcall
 local luaObjectLines = {
@@ -150,7 +151,7 @@ return {
       path_resolution_modifier = {},
     },
     LuaAccumulatorControlBehavior = {
-      type = {readOnly = true},
+      type = {readOnly = true, enum = invert(defines.control_behavior.type,"defines.control_behavior.type.")},
       entity = {readOnly = true},
       output_signal = {},
     },
@@ -170,7 +171,7 @@ return {
       bonus_gui_order = {readOnly = true},
     },
     LuaArithmeticCombinatorControlBehavior = {
-      type = {readOnly = true},
+      type = {readOnly = true, enum = invert(defines.control_behavior.type,"defines.control_behavior.type.")},
       entity = {readOnly = true},
       signals_last_tick = {readOnly = true},
       parameters = {readOnly = true},
@@ -212,8 +213,37 @@ return {
     },
     LuaCircuitNetwork = {
       entity = {readOnly = true},
-      wire_type = {readOnly = true},
-      circuit_connector_id = {readOnly = true},
+      wire_type = {readOnly = true, enum = invert(defines.wire_type,"defines.wire_type.")},
+      circuit_connector_id = {readOnly = true, enum = (function()
+        local combinator = invert(defines.circuit_connector_id,"defines.circuit_connector_id.",function(k,v) return (not not string.match(k,"^combinator")) end)
+        local netnames = {
+          ["accumulator"] = {[defines.circuit_connector_id.accumulator] = "defines.circuit_connector_id.accumulator"},
+          ["container"] = {[defines.circuit_connector_id.container] = "defines.circuit_connector_id.container"},
+          ["logistic-container"] = {[defines.circuit_connector_id.container] = "defines.circuit_connector_id.container"},
+          ["programmable-speaker"] = {[defines.circuit_connector_id.programmable_speaker] = "defines.circuit_connector_id.programmable_speaker"},
+          ["rail-signal"] = {[defines.circuit_connector_id.rail_signal] = "defines.circuit_connector_id.rail_signal"},
+          ["rail-chain-signal"] = {[defines.circuit_connector_id.rail_chain_signal] = "defines.circuit_connector_id.rail_chain_signal"},
+          ["roboport"] = {[defines.circuit_connector_id.roboport] = "defines.circuit_connector_id.roboport"},
+          ["storage-tank"] = {[defines.circuit_connector_id.storage_tank] = "defines.circuit_connector_id.storage_tank"},
+          ["wall"] = {[defines.circuit_connector_id.wall] = "defines.circuit_connector_id.wall"},
+          ["electric-pole"] = {[defines.circuit_connector_id.electric_pole] = "defines.circuit_connector_id.electric_pole"},
+          ["inserter"] = {[defines.circuit_connector_id.inserter] = "defines.circuit_connector_id.inserter"},
+          ["lamp"] = {[defines.circuit_connector_id.lamp] = "defines.circuit_connector_id.lamp"},
+          ["pump"] = {[defines.circuit_connector_id.pump] = "defines.circuit_connector_id.pump"},
+          ["ofshore-pump"] = {[defines.circuit_connector_id.offshore_pump] = "defines.circuit_connector_id.ofshore_pump"},
+
+          ["constant-combinator"] = {[defines.circuit_connector_id.constant_combinator] = "defines.circuit_connector_id.constant_combinator"},
+
+          ["decider-combinator"] = combinator,
+          ["arithmetic-combinator"] = combinator,
+        }
+        return function(network,id)
+          local names = netnames[network.entity.type]
+          if names then
+            return names[id]
+          end
+        end
+      end)()},
       signals = {readOnly = true},
       network_id = {readOnly = true},
       connected_circuit_count = {readOnly = true},
@@ -223,14 +253,14 @@ return {
       game_commands = {readOnly = true},
     },
     LuaConstantCombinatorControlBehavior = {
-      type = {readOnly = true},
+      type = {readOnly = true, enum = invert(defines.control_behavior.type,"defines.control_behavior.type.")},
       entity = {readOnly = true},
       parameters = {readOnly = true},
       enabled = {},
       signals_count = {readOnly = true},
     },
     LuaContainerControlBehavior = {
-      type = {readOnly = true},
+      type = {readOnly = true, enum = invert(defines.control_behavior.type,"defines.control_behavior.type.")},
       entity = {readOnly = true},
     },
     LuaCustomChartTag = {
@@ -261,7 +291,7 @@ return {
       hidden = {readOnly = true},
     },
     LuaDeciderCombinatorControlBehavior = {
-      type = {readOnly = true},
+      type = {readOnly = true, enum = invert(defines.control_behavior.type,"defines.control_behavior.type.")},
       entity = {readOnly = true},
       signals_last_tick = {readOnly = true},
       parameters = {readOnly = true},
@@ -330,7 +360,7 @@ return {
       character_personal_logistic_requests_enabled = {},
       vehicle_logistic_requests_enabled = {},
       auto_trash_filters = {},
-      opened_gui_type = {readOnly = true},
+      opened_gui_type = {readOnly = true, enum = invert(defines.gui_type,"defines.gui_type.")},
       build_distance = {readOnly = true},
       drop_item_distance = {readOnly = true},
       reach_distance = {readOnly = true},
@@ -355,7 +385,7 @@ return {
       rotatable = {},
       operable = {},
       health = {},
-      direction = {},
+      direction = {enum = invert(defines.direction,"defines.direction.")},
       supports_direction = {readOnly = true},
       orientation = {},
       cliff_orientation = {readOnly = true},
@@ -389,8 +419,8 @@ return {
       time_to_live = {},
       color = {},
       text = {},
-      signal_state = {readOnly = true},
-      chain_signal_state = {readOnly = true},
+      signal_state = {readOnly = true, enum = invert(defines.signal_state,"defines.signal_state.")},
+      chain_signal_state = {readOnly = true, enum = invert(defines.chain_signal_state,"defines.chain_signal_state.")},
       to_be_looted = {},
       crafting_speed = {readOnly = true},
       crafting_progress = {},
@@ -475,7 +505,7 @@ return {
       ai_settings = {readOnly = true},
       highlight_box_type = {},
       highlight_box_blink_interval = {},
-      status = {readOnly = true},
+      status = {readOnly = true, enum = invert(defines.entity_status,"defines.entity_status.")},
       enable_logistics_while_moving = {},
       render_player = {},
       render_to_forces = {},
@@ -887,7 +917,7 @@ return {
       players = {readOnly = true},
       map_settings = {readOnly = true},
       difficulty_settings = {readOnly = true},
-      difficulty = {readOnly = true},
+      difficulty = {readOnly = true, enum = invert(defines.difficulty,"defines.difficulty.")},
       forces = {readOnly = true},
       entity_prototypes = {readOnly = true},
       item_prototypes = {readOnly = true},
@@ -959,7 +989,7 @@ return {
       show_shortcut_bar = {},
     },
     LuaGenericOnOffControlBehavior = {
-      type = {readOnly = true},
+      type = {readOnly = true, enum = invert(defines.control_behavior.type,"defines.control_behavior.type.")},
       entity = {readOnly = true},
       disabled = {readOnly = true},
       circuit_condition = {},
@@ -1056,15 +1086,15 @@ return {
       right_label_tooltip = {},
     },
     LuaInserterControlBehavior = {
-      type = {readOnly = true},
+      type = {readOnly = true, enum = invert(defines.control_behavior.type,"defines.control_behavior.type.")},
       entity = {readOnly = true},
       disabled = {readOnly = true},
       circuit_condition = {},
       logistic_condition = {},
       connect_to_logistic_network = {},
       circuit_read_hand_contents = {},
-      circuit_mode_of_operation = {},
-      circuit_hand_read_mode = {},
+      circuit_mode_of_operation = {enum = invert(defines.control_behavior.inserter.circuit_mode_of_operation,"defines.control_behavior.inserter.circuit_mode_of_operation.")},
+      circuit_hand_read_mode = {enum = invert(defines.control_behavior.inserter.hand_read_mode,"defines.control_behavior.inserter.hand_read_mode.")},
       circuit_set_stack_size = {},
       circuit_stack_control_signal = {},
     },
@@ -1082,7 +1112,90 @@ return {
       connections = {readOnly = true},
     },
     LuaInventory = {
-      index = {readOnly = true},
+      index = {readOnly = true, enum =(function()
+        local burner = {
+          [defines.inventory.fuel] = "defines.inventory.fuel",
+          [defines.inventory.burnt_result] = "defines.inventory.burnt_result",
+        }
+        local function with(super,t) return setmetatable(t,{__index = super}) end
+
+        local chest = {
+          [defines.inventory.chest] = "defines.inventory.chest",
+        }
+
+        local assembler = with(burner,invert(defines.inventory,"defines.inventory.",function(k,v) return not not string.match(k,"^assembling_machine_") end))
+
+        local character = invert(defines.inventory,"defines.inventory.",function(k,v) return (not not string.match(k,"^character_")) and k ~= "character_corpse" end)
+        local robot = invert(defines.inventory,"defines.inventory.",function(k,v) return (not not string.match(k,"^robot_")) end)
+
+        local invname = {
+          burner = burner,
+          item = invert(defines.inventory,"defines.inventory.",function(k,v) return not not string.match(k,"^item_") end),
+          player = {
+            [defines.controllers.character] = character,
+            [defines.controllers.god] =
+              invert(defines.inventory,"defines.inventory.",function(k,v) return not not string.match(k,"^god_") end),
+            [defines.controllers.editor] =
+              invert(defines.inventory,"defines.inventory.",function(k,v) return not not string.match(k,"^editor_") end),
+          },
+          entity = {
+            ["container"]=chest,
+            ["logistic-container"]=chest,
+            ["cargo-wagon"]={ [defines.inventory.cargo_wagon] = "defines.inventory.cargo_wagon" },
+            ["rocket-silo-rocket"]={ [defines.inventory.cargo_wagon] = "defines.inventory.rocket" },
+
+            ["construction-robot"]=robot,
+            ["logistic-robot"]=robot,
+
+            ["ammo-turret"]=invert(defines.inventory,"defines.inventory.",function(k,v) return not not string.match(k,"^turret_") end),
+            ["artillery-turret"]=invert(defines.inventory,"defines.inventory.",function(k,v) return not not string.match(k,"^artillery_turret_") end),
+            ["artillery-wagon"]=invert(defines.inventory,"defines.inventory.",function(k,v) return not not string.match(k,"^artillery_turret_") end),
+            ["roboport"]=invert(defines.inventory,"defines.inventory.",function(k,v) return not not string.match(k,"^roboport_") end),
+            ["beacon"]=invert(defines.inventory,"defines.inventory.",function(k,v) return not not string.match(k,"^beacon_") end),
+            ["character-corpse"]=invert(defines.inventory,"defines.inventory.",function(k,v) return not not string.match(k,"^character_corpse_") end),
+
+            ["furnace"]=with(burner,invert(defines.inventory,"defines.inventory.",function(k,v) return not not string.match(k,"^furnace_") end)),
+            ["assembling-machine"]=assembler,
+            ["mining-drill"]=with(burner,invert(defines.inventory,"defines.inventory.",function(k,v) return not not string.match(k,"^mining_drill_") end)),
+            ["lab"]=with(burner,invert(defines.inventory,"defines.inventory.",function(k,v) return not not string.match(k,"^lab_") end)),
+            ["car"]=with(burner,invert(defines.inventory,"defines.inventory.",function(k,v) return not not string.match(k,"^car_") end)),
+            ["spider-vehicle"]=with(burner,invert(defines.inventory,"defines.inventory.",function(k,v) return not not string.match(k,"^spider_") end)),
+            ["rocket-silo"]=with(assembler,invert(defines.inventory,"defines.inventory.",function(k,v) return not not string.match(k,"^rocket_silo_") end)),
+
+          },
+        }
+        return function(inv,index)
+          local owner = inv.player_owner
+          if owner then
+            -- check if player is character/god/editor
+            local names = invname.player[owner.controller_type]
+            if names then
+              return names[index]
+            end
+            return
+          end
+          owner = inv.equipment_owner
+          if owner then
+            -- burner inside equipment
+            return invname.burner[index]
+          end
+          owner = inv.entity_owner
+          if owner then
+            -- check entity type
+            local names = invname.entity[owner.type]
+            if names then
+              return names[index]
+            end
+            return
+          end
+          owner = inv.mod_owner
+          if owner then
+            return nil
+          end
+          local names = invname.item
+          return names[index]
+        end
+      end)()},
       entity_owner = {readOnly = true},
       player_owner = {readOnly = true},
       equipment_owner = {readOnly = true},
@@ -1187,9 +1300,9 @@ return {
       custom_description = {},
       entity_filters = {},
       tile_filters = {},
-      entity_filter_mode = {},
-      tile_filter_mode = {},
-      tile_selection_mode = {},
+      entity_filter_mode = {enum = invert(defines.deconstruction_item.entity_filter_mode,"defines.deconstruction_item.entity_filter_mode.")},
+      tile_filter_mode = {enum = invert(defines.deconstruction_item.tile_filter_mode,"defines.deconstruction_item.tile_filter_mode.")},
+      tile_selection_mode = {enum = invert(defines.deconstruction_item.tile_selection_mode,"defines.deconstruction_item.tile_selection_mode.")},
       trees_and_rocks_only = {},
       entity_filter_count = {readOnly = true},
       tile_filter_count = {readOnly = true},
@@ -1212,7 +1325,7 @@ return {
       is_upgrade_item = {readOnly = true},
     },
     LuaLampControlBehavior = {
-      type = {readOnly = true},
+      type = {readOnly = true, enum = invert(defines.control_behavior.type,"defines.control_behavior.type.")},
       entity = {readOnly = true},
       disabled = {readOnly = true},
       circuit_condition = {},
@@ -1239,9 +1352,9 @@ return {
       to_charge_robots = {readOnly = true},
     },
     LuaLogisticContainerControlBehavior = {
-      type = {readOnly = true},
+      type = {readOnly = true, enum = invert(defines.control_behavior.type,"defines.control_behavior.type.")},
       entity = {readOnly = true},
-      circuit_mode_of_operation = {},
+      circuit_mode_of_operation = {enum = invert(defines.control_behavior.logistic_container.circuit_mode_of_operation,"defines.control_behavior.logistic_container.circuit_mode_of_operation.")},
     },
     LuaLogisticNetwork = {
       force = {readOnly = true},
@@ -1271,7 +1384,7 @@ return {
       logistic_network = {readOnly = true},
       logistic_member_index = {readOnly = true},
       filters = {readOnly = true},
-      mode = {readOnly = true},
+      mode = {readOnly = true, enum = invert(defines.logistic_mode,"defines.logistic_mode.")},
       force = {readOnly = true},
       targeted_items_pickup = {readOnly = true},
       targeted_items_deliver = {readOnly = true},
@@ -1386,7 +1499,7 @@ return {
     },
 
     LuaMiningDrillControlBehavior = {
-      type = {readOnly = true},
+      type = {readOnly = true, enum = invert(defines.control_behavior.type,"defines.control_behavior.type.")},
       entity = {readOnly = true},
       disabled = {readOnly = true},
       circuit_condition = {},
@@ -1394,7 +1507,7 @@ return {
       connect_to_logistic_network = {},
       circuit_enable_disable = {},
       circuit_read_resources = {},
-      resource_read_mode = {},
+      resource_read_mode = {enum = invert(defines.control_behavior.mining_drill.resource_read_mode,"defines.control_behavior.mining_drill.resource_read_mode.")},
       resource_read_targets = {readOnly = true},
     },
     LuaModSettingPrototype = {
@@ -1493,7 +1606,7 @@ return {
       character_personal_logistic_requests_enabled = {},
       vehicle_logistic_requests_enabled = {},
       auto_trash_filters = {},
-      opened_gui_type = {readOnly = true},
+      opened_gui_type = {readOnly = true, enum = invert(defines.gui_type,"defines.gui_type.")},
       build_distance = {readOnly = true},
       drop_item_distance = {readOnly = true},
       reach_distance = {readOnly = true},
@@ -1508,8 +1621,8 @@ return {
       index = {readOnly = true},
       gui = {readOnly = true},
       opened_self = {readOnly = true},
-      controller_type = {readOnly = true},
-      stashed_controller_type = {readOnly = true},
+      controller_type = {readOnly = true, enum = invert(defines.controllers,"defines.controllers.")},
+      stashed_controller_type = {readOnly = true, enum = invert(defines.controllers,"defines.controllers.")},
       game_view_settings = {},
       minimap_enabled = {},
       color = {},
@@ -1528,7 +1641,7 @@ return {
       display_resolution = {readOnly = true},
       display_scale = {readOnly = true},
       blueprint_to_setup = {readOnly = true},
-      render_mode = {readOnly = true},
+      render_mode = {readOnly = true, enum = invert(defines.render_mode,"defines.render_mode.")},
       spectator = {},
       remove_unfiltered_items = {},
       infinity_inventory_filters = {},
@@ -1539,13 +1652,13 @@ return {
       ["<translated>"] = {thisTranslated = true},
     },
     LuaProgrammableSpeakerControlBehavior = {
-      type = {readOnly = true},
+      type = {readOnly = true, enum = invert(defines.control_behavior.type,"defines.control_behavior.type.")},
       entity = {readOnly = true},
       circuit_parameters = {},
       circuit_condition = {},
     },
     LuaRailChainSignalControlBehavior = {
-      type = {readOnly = true},
+      type = {readOnly = true, enum = invert(defines.control_behavior.type,"defines.control_behavior.type.")},
       entity = {readOnly = true},
       red_signal = {},
       orange_signal = {},
@@ -1560,7 +1673,7 @@ return {
       rails = {readOnly = true},
     },
     LuaRailSignalControlBehavior = {
-      type = {readOnly = true},
+      type = {readOnly = true, enum = invert(defines.control_behavior.type,"defines.control_behavior.type.")},
       entity = {readOnly = true},
       red_signal = {},
       orange_signal = {},
@@ -1629,7 +1742,7 @@ return {
       localised_description = {readOnly = true},
     },
     LuaRoboportControlBehavior = {
-      type = {readOnly = true},
+      type = {readOnly = true, enum = invert(defines.control_behavior.type,"defines.control_behavior.type.")},
       entity = {readOnly = true},
       read_logistics = {},
       read_robot_stats = {},
@@ -1655,7 +1768,7 @@ return {
       associated_control_input = {readOnly = true},
     },
     LuaStorageTankControlBehavior = {
-      type = {readOnly = true},
+      type = {readOnly = true, enum = invert(defines.control_behavior.type,"defines.control_behavior.type.")},
       entity = {readOnly = true},
     },
     LuaStyle = {
@@ -1687,7 +1800,7 @@ return {
       vertically_stretchable = {},
       horizontally_squashable = {},
       vertically_squashable = {},
-      rich_text_setting = {},
+      rich_text_setting = {enum = invert(defines.rich_text_setting,"defines.rich_text_setting.")},
       hovered_font_color = {},
       clicked_font_color = {},
       disabled_font_color = {},
@@ -1819,11 +1932,11 @@ return {
       cargo_wagons = {readOnly = true},
       fluid_wagons = {readOnly = true},
       schedule = {},
-      state = {readOnly = true},
+      state = {readOnly = true, enum = invert(defines.train_state,"defines.train_state.")},
       front_rail = {readOnly = true},
       back_rail = {readOnly = true},
-      rail_direction_from_front_rail = {readOnly = true},
-      rail_direction_from_back_rail = {readOnly = true},
+      rail_direction_from_front_rail = {readOnly = true, enum = invert(defines.rail_direction,"defines.rail_direction.")},
+      rail_direction_from_back_rail = {readOnly = true, enum = invert(defines.rail_direction,"defines.rail_direction.")},
       front_stock = {readOnly = true},
       back_stock = {readOnly = true},
       station = {readOnly = true},
@@ -1839,7 +1952,7 @@ return {
       signal = {readOnly = true},
     },
     LuaTrainStopControlBehavior = {
-      type = {readOnly = true},
+      type = {readOnly = true, enum = invert(defines.control_behavior.type,"defines.control_behavior.type.")},
       entity = {readOnly = true},
       disabled = {readOnly = true},
       circuit_condition = {},
@@ -1856,7 +1969,7 @@ return {
       trains_limit_signal = {},
     },
     LuaTransportBeltControlBehavior = {
-      type = {readOnly = true},
+      type = {readOnly = true, enum = invert(defines.control_behavior.type,"defines.control_behavior.type.")},
       entity = {readOnly = true},
       disabled = {readOnly = true},
       circuit_condition = {},
@@ -1864,7 +1977,7 @@ return {
       connect_to_logistic_network = {},
       enable_disable = {},
       read_contents = {},
-      read_contents_mode = {},
+      read_contents_mode = {enum = invert(defines.control_behavior.transport_belt.content_read_mode,"defines.control_behavior.transport_belt.content_read_mode.")},
     },
     LuaTransportLine = {
       owner = {readOnly = true},
@@ -1895,7 +2008,7 @@ return {
     LuaUnitGroup = {
       members = {readOnly = true},
       position = {readOnly = true},
-      state = {readOnly = true},
+      state = {readOnly = true, enum = invert(defines.group_state,"defines.group_state.")},
       force = {readOnly = true},
       surface = {readOnly = true},
       group_number = {readOnly = true},
@@ -1917,7 +2030,7 @@ return {
       render_no_power_icon = {readOnly = true},
     },
     LuaWallControlBehavior = {
-      type = {readOnly = true},
+      type = {readOnly = true, enum = invert(defines.control_behavior.type,"defines.control_behavior.type.")},
       entity = {readOnly = true},
       circuit_condition = {},
       open_gate = {},